Gemini-Text-Summarization-App/requirements.txt
protobuf==3.20.3
streamlit==1.16.0
genai==2.1.0
creds==1.0.15
google-generativeai==0.3.2